deprecate voxel encoder and make n dimensional instead (enabling 2D etc..)
